+++ modules/kubernetes/03-k8s_controllers.tf
@@ -17,7 +17,7 @@ data "template_file" "bootstrap_node_k8s_controllers" {
 resource "aws_launch_configuration" "k8s_controllers_launch_configuration" {
   image_id                    = var.ami_id_controller != "" ? var.ami_id_controller : data.aws_ami.ami_dynamic.id
   instance_type               = var.ec2_k8s_controllers_instance_type
-  key_name                    = var.ec2_key_name
+  key_name                    = var.enable_ssm_access_to_nodes ? null : var.ec2_key_name
   user_data                   = data.template_file.bootstrap_node_k8s_controllers.rendered
   iam_instance_profile        = aws_iam_instance_profile.k8s_instance_profile.id
   associate_public_ip_address = false