+++ modules/kubernetes/05-role.tf
@@ -43,7 +43,7 @@ resource "aws_iam_role" "k8s_assume_role" {
     {
       "Action": "sts:AssumeRole",
       "Principal": {
-        "Service": "ec2.amazonaws.com"
+        "Service": "ec2.${data.aws_partition.current.dns_suffix}"
       },
       "Effect": "Allow",
       "Sid": ""